+++ src/Dithering/FloydSteinbergDithering.cs
@@ -6,15 +6,13 @@
  * Licensed under the MIT License. See LICENSE.txt for the full text.
  */
 
-using Cyotek.DitheringTest.Helpers;
-
-namespace Cyotek.DitheringTest.Dithering
+namespace Cyotek.Drawing.Imaging.ColorReduction
 {
   public sealed class FloydSteinbergDithering : IErrorDiffusion
   {
-    #region IDither Interface
+    #region IErrorDiffusion Interface
 
-    void IErrorDiffusion.Diffuse(ArgbColor[] original, ArgbColor originalPixel, ArgbColor transformedPixel, int x, int y, int width, int height)
+    void IErrorDiffusion.Diffuse(ArgbColor[] data, ArgbColor original, ArgbColor transformed, int x, int y, int width, int height)
     {
       ArgbColor offsetPixel;
       int redError;
@@ -37,19 +35,19 @@ namespace Cyotek.DitheringTest.Dithering
         */
 
       index = y * width + x;
-      redError = originalPixel.R - transformedPixel.R;
-      blueError = originalPixel.G - transformedPixel.G;
-      greenError = originalPixel.B - transformedPixel.B;
+      redError = original.R - transformed.R;
+      blueError = original.G - transformed.G;
+      greenError = original.B - transformed.B;
 
       if (x + 1 < width)
       {
         // right
         offsetIndex = index + 1;
-        offsetPixel = original[offsetIndex];
+        offsetPixel = data[offsetIndex];
         offsetPixel.R = (offsetPixel.R + ((redError * 7) >> 4)).ToByte();
         offsetPixel.G = (offsetPixel.G + ((greenError * 7) >> 4)).ToByte();
         offsetPixel.B = (offsetPixel.B + ((blueError * 7) >> 4)).ToByte();
-        original[offsetIndex] = offsetPixel;
+        data[offsetIndex] = offsetPixel;
       }
 
       if (y + 1 < height)
@@ -58,30 +56,30 @@ namespace Cyotek.DitheringTest.Dithering
         {
           // left and down
           offsetIndex = index + width - 1;
-          offsetPixel = original[offsetIndex];
+          offsetPixel = data[offsetIndex];
           offsetPixel.R = (offsetPixel.R + ((redError * 3) >> 4)).ToByte();
           offsetPixel.G = (offsetPixel.G + ((greenError * 3) >> 4)).ToByte();
           offsetPixel.B = (offsetPixel.B + ((blueError * 3) >> 4)).ToByte();
-          original[offsetIndex] = offsetPixel;
+          data[offsetIndex] = offsetPixel;
         }
 
         // down
         offsetIndex = index + width;
-        offsetPixel = original[offsetIndex];
+        offsetPixel = data[offsetIndex];
         offsetPixel.R = (offsetPixel.R + ((redError * 5) >> 4)).ToByte();
         offsetPixel.G = (offsetPixel.G + ((greenError * 5) >> 4)).ToByte();
         offsetPixel.B = (offsetPixel.B + ((blueError * 5) >> 4)).ToByte();
-        original[offsetIndex] = offsetPixel;
+        data[offsetIndex] = offsetPixel;
 
         if (x + 1 < width)
         {
           // right and down
           offsetIndex = index + width + 1;
-          offsetPixel = original[offsetIndex];
+          offsetPixel = data[offsetIndex];
           offsetPixel.R = (offsetPixel.R + ((redError * 1) >> 4)).ToByte();
           offsetPixel.G = (offsetPixel.G + ((greenError * 1) >> 4)).ToByte();
           offsetPixel.B = (offsetPixel.B + ((blueError * 1) >> 4)).ToByte();
-          original[offsetIndex] = offsetPixel;
+          data[offsetIndex] = offsetPixel;
         }
       }
     }