+++ src/Dithering/JarvisJudiceNinkeDithering.cs
@@ -7,29 +7,33 @@
  */
 
 /*
- * Burkes Dithering
+ * Jarvis, Judice & Ninke Dithering
  * http://www.efg2.com/Lab/Library/ImageProcessing/DHALF.TXT
  *
- *                  *  8/32 4/32
- *      2/32 4/32 8/32 4/32 2/32
+ *                  *  8/48 5/48
+ *      3/48 5/48 7/48 5/48 3/48
+ *      1/48 3/48 5/48 3/48 1/48
  */
 
 namespace Cyotek.Drawing.Imaging.ColorReduction
 {
-  public sealed class BurksDithering : ErrorDiffusionDithering
+  public sealed class JarvisJudiceNinkeDithering : ErrorDiffusionDithering
   {
     #region Constructors
 
-    public BurksDithering()
+    public JarvisJudiceNinkeDithering()
       : base(new byte[,]
              {
                {
-                 0, 0, 0, 8, 4
+                 0, 0, 0, 7, 5
                },
                {
-                 2, 4, 8, 4, 2
+                 3, 5, 7, 5, 3
+               },
+               {
+                 1, 3, 5, 3, 1
                }
-             }, 5, true)
+             }, 48, false)
     { }
 
     #endregion