+++ CHANGELOG.md
@@ -1,2 +1,13 @@
-## 06Jun2015
-* Initial Release
+Change Log
+==========
+
+1.0.1.0
+-------
+* Namespaces adjusted to account for library linking
+* Added new base class to handle error diffusion algorithms by simply specifying a matrix
+* Added Jarvis, Judice & Ninke, Sierra3, Sierra2, SierraLite, Stucki and Atkinson algorithms
+* Added Random dithering
+
+1.0.0.0
+-------
+* Initial release